fix: Corrected segfault at program shutdown
Task 0 involved debugging a segfault at program shutdown. This was
caused by two things:

    1) The ChatLogic attempting to delete the ChatBot even though the
       the handle is listed as "(not owned)".

    2) The GraphNodes not initializing the ChatBot handle to nullptr,
       which causes problems when the node tries to delete garbage.

Having the GraphNode own the ChatBot seems more error prone than
allowing the ChatLogic to manage it, but unfortunately that part
of the ChatLogic code lies outside of the STUDENT CODE block.

Note also that the ChatBot can be leaked if there are problems
constructing the graph and the bot is never transferred to a node.